+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.0.1] - 08-04-2025
+### Added
+- Unit tests for Variables (BoolVariable, FloatVariable, IntVariable, StringVariable).
+- Unit tests for References (BoolReference, FloatReference, IntReference, StringReference).
+
 ## [1.0.0] - 07-04-2025
 
 ### First release